+++ prefect_snowflake/database.py
@@ -160,13 +160,19 @@ class SnowflakeConnector(DatabaseBlock):
         if self._unique_cursors is None:
             self._unique_cursors = {}
 
-    def _get_cursor(self, inputs: Dict[str, Any]) -> Tuple[bool, SnowflakeCursor]:
+    def _get_cursor(
+        self,
+        inputs: Dict[str, Any],
+        cursor_class: type[SnowflakeCursor] = SnowflakeCursor,
+    ) -> Tuple[bool, SnowflakeCursor]:
         """
         Get a Snowflake cursor.
 
         Args:
             inputs: The inputs to generate a unique hash, used to decide
                 whether a new cursor should be used.
+            cursor_class: The class of the cursor to use when creating a
+                Snowflake cursor.
 
         Returns:
             Whether a cursor is new and a Snowflake cursor.
@@ -181,7 +187,7 @@ class SnowflakeConnector(DatabaseBlock):
                 "please open an issue with a reproducible example."
             )
         if input_hash not in self._unique_cursors.keys():
-            new_cursor = self._connection.cursor()
+            new_cursor = self._connection.cursor(cursor_class=cursor_class)
             self._unique_cursors[input_hash] = new_cursor
             return True, new_cursor
         else:
@@ -251,6 +257,7 @@ class SnowflakeConnector(DatabaseBlock):
         self,
         operation: str,
         parameters: Optional[Dict[str, Any]] = None,
+        cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
         **execute_kwargs: Dict[str, Any],
     ) -> Tuple[Any]:
         """
@@ -263,6 +270,7 @@ class SnowflakeConnector(DatabaseBlock):
         Args:
             operation: The SQL query or other operation to be executed.
             parameters: The parameters for the operation.
+            cursor_class: The class of the cursor to use when creating a Snowflake cursor.
             **execute_kwargs: Additional options to pass to `cursor.execute_async`.
 
         Returns:
@@ -298,7 +306,7 @@ class SnowflakeConnector(DatabaseBlock):
             params=parameters,
             **execute_kwargs,
         )
-        new, cursor = self._get_cursor(inputs)
+        new, cursor = self._get_cursor(inputs, cursor_class=cursor_class)
         if new:
             await self._execute_async(cursor, inputs)
         self.logger.debug("Preparing to fetch a row.")
@@ -311,6 +319,7 @@ class SnowflakeConnector(DatabaseBlock):
         operation: str,
         parameters: Optional[Dict[str, Any]] = None,
         size: Optional[int] = None,
+        cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
         **execute_kwargs: Dict[str, Any],
     ) -> List[Tuple[Any]]:
         """
@@ -325,6 +334,7 @@ class SnowflakeConnector(DatabaseBlock):
             parameters: The parameters for the operation.
             size: The number of results to return; if None or 0, uses the value of
                 `fetch_size` configured on the block.
+            cursor_class: The class of the cursor to use when creating a Snowflake cursor.
             **execute_kwargs: Additional options to pass to `cursor.execute_async`.
 
         Returns:
@@ -368,7 +378,7 @@ class SnowflakeConnector(DatabaseBlock):
             params=parameters,
             **execute_kwargs,
         )
-        new, cursor = self._get_cursor(inputs)
+        new, cursor = self._get_cursor(inputs, cursor_class)
         if new:
             await self._execute_async(cursor, inputs)
         size = size or self.fetch_size
@@ -381,6 +391,7 @@ class SnowflakeConnector(DatabaseBlock):
         self,
         operation: str,
         parameters: Optional[Dict[str, Any]] = None,
+        cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
         **execute_kwargs: Dict[str, Any],
     ) -> List[Tuple[Any]]:
         """
@@ -393,6 +404,7 @@ class SnowflakeConnector(DatabaseBlock):
         Args:
             operation: The SQL query or other operation to be executed.
             parameters: The parameters for the operation.
+            cursor_class: The class of the cursor to use when creating a Snowflake cursor.
             **execute_kwargs: Additional options to pass to `cursor.execute_async`.
 
         Returns:
@@ -429,7 +441,7 @@ class SnowflakeConnector(DatabaseBlock):
             params=parameters,
             **execute_kwargs,
         )
-        new, cursor = self._get_cursor(inputs)
+        new, cursor = self._get_cursor(inputs, cursor_class)
         if new:
             await self._execute_async(cursor, inputs)
         self.logger.debug("Preparing to fetch all rows.")
@@ -441,6 +453,7 @@ class SnowflakeConnector(DatabaseBlock):
         self,
         operation: str,
         parameters: Optional[Dict[str, Any]] = None,
+        cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
         **execute_kwargs: Dict[str, Any],
     ) -> None:
         """
@@ -452,6 +465,7 @@ class SnowflakeConnector(DatabaseBlock):
         Args:
             operation: The SQL query or other operation to be executed.
             parameters: The parameters for the operation.
+            cursor_class: The class of the cursor to use when creating a Snowflake cursor.
             **execute_kwargs: Additional options to pass to `cursor.execute_async`.
 
         Examples:
@@ -472,7 +486,7 @@ class SnowflakeConnector(DatabaseBlock):
             params=parameters,
             **execute_kwargs,
         )
-        with self._connection.cursor() as cursor:
+        with self._connection.cursor(cursor_class) as cursor:
             await run_sync_in_worker_thread(cursor.execute, **inputs)
         self.logger.info(f"Executed the operation, {operation!r}.")
 
@@ -566,7 +580,7 @@ async def snowflake_query(
     query: str,
     snowflake_connector: SnowflakeConnector,
     params: Union[Tuple[Any], Dict[str, Any]] = None,
-    cursor_type: SnowflakeCursor = SnowflakeCursor,
+    cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
     poll_frequency_seconds: int = 1,
 ) -> List[Tuple[Any]]:
     """
@@ -576,7 +590,7 @@ async def snowflake_query(
         query: The query to execute against the database.
         params: The params to replace the placeholders in the query.
         snowflake_connector: The credentials to use to authenticate.
-        cursor_type: The type of database cursor to use for the query.
+        cursor_class: The type of database cursor to use for the query.
         poll_frequency_seconds: Number of seconds to wait in between checks for
             run completion.
 
@@ -616,7 +630,7 @@ async def snowflake_query(
     """
     # context manager automatically rolls back failed transactions and closes
     with snowflake_connector.get_connection() as connection:
-        with connection.cursor(cursor_type) as cursor:
+        with connection.cursor(cursor_class) as cursor:
             response = cursor.execute_async(query, params=params)
             query_id = response["queryId"]
             while connection.is_still_running(
@@ -633,7 +647,7 @@ async def snowflake_multiquery(
     queries: List[str],
     snowflake_connector: SnowflakeConnector,
     params: Union[Tuple[Any], Dict[str, Any]] = None,
-    cursor_type: SnowflakeCursor = SnowflakeCursor,
+    cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
     as_transaction: bool = False,
     return_transaction_control_results: bool = False,
     poll_frequency_seconds: int = 1,
@@ -646,7 +660,7 @@ async def snowflake_multiquery(
         queries: The list of queries to execute against the database.
         params: The params to replace the placeholders in the query.
         snowflake_connector: The credentials to use to authenticate.
-        cursor_type: The type of database cursor to use for the query.
+        cursor_class: The type of database cursor to use for the query.
         as_transaction: If True, queries are executed in a transaction.
         return_transaction_control_results: Determines if the results of queries
             controlling the transaction (BEGIN/COMMIT) should be returned.
@@ -693,7 +707,7 @@ async def snowflake_multiquery(
             queries.insert(0, BEGIN_TRANSACTION_STATEMENT)
             queries.append(END_TRANSACTION_STATEMENT)
 
-        with connection.cursor(cursor_type) as cursor:
+        with connection.cursor(cursor_class) as cursor:
             results = []
             for query in queries:
                 response = cursor.execute_async(query, params=params)
@@ -718,7 +732,7 @@ async def snowflake_query_sync(
     query: str,
     snowflake_connector: SnowflakeConnector,
     params: Union[Tuple[Any], Dict[str, Any]] = None,
-    cursor_type: SnowflakeCursor = SnowflakeCursor,
+    cursor_class: Optional[type[SnowflakeCursor]] = SnowflakeCursor,
 ) -> List[Tuple[Any]]:
     """
     Executes a query in sync mode against a Snowflake database.
@@ -727,7 +741,7 @@ async def snowflake_query_sync(
         query: The query to execute against the database.
         params: The params to replace the placeholders in the query.
         snowflake_connector: The credentials to use to authenticate.
-        cursor_type: The type of database cursor to use for the query.
+        cursor_class: The type of database cursor to use for the query.
 
     Returns:
         The output of `response.fetchall()`.
@@ -764,7 +778,7 @@ async def snowflake_query_sync(
     """
     # context manager automatically rolls back failed transactions and closes
     with snowflake_connector.get_connection() as connection:
-        with connection.cursor(cursor_type) as cursor:
+        with connection.cursor(cursor_class) as cursor:
             cursor.execute(query, params=params)
             result = cursor.fetchall()
     return result